Moved MemoryCache to different catalog
