+++ GenericTableAPI/Extensions/MemoryCacheExtension.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.Extensions.Caching.Memory;
 
-namespace GenericTableAPI.Utilities
+namespace GenericTableAPI.Extensions
 {
-    public static class MemoryCacheHandler
+    public static class MemoryCacheExtension
     {
         public static int CacheDurationSeconds { get; set; } = 0;
 
         public static void SetCache(this IMemoryCache memoryCache, string cache, object? obj)
         {
-            if(CacheDurationSeconds <= 0)
+            if (CacheDurationSeconds <= 0)
                 return;
             memoryCache.Set(cache, obj, new MemoryCacheEntryOptions().SetAbsoluteExpiration(TimeSpan.FromSeconds(CacheDurationSeconds)));
         }
